+++ debug/site.js
@@ -5,23 +5,23 @@ var llmr = require('../'),
     style_json = require('./style-streets');
 
 if (typeof document !== 'undefined') {
-    var map = new llmr.Map({
-        container: document.getElementById('map'),
+    var map = window.map = new llmr.Map({
+        container: 'map',
         sources: {
             "mapbox streets": {
                 type: 'vector',
-                urls: ['http://a.gl-api-us-east-1.tilestream.net/v3/mapbox.mapbox-streets-v4/{z}/{x}/{y}.gl.pbf'],
+                url: 'http://{s}.gl-api-us-east-1.tilestream.net/v3/mapbox.mapbox-streets-v4/{z}/{x}/{y}.gl.pbf',
                 // urls: ['http://api.tiles.mapbox.com/v3/mapbox.mapbox-streets-v4/{z}/{x}/{y}.vector.pbf'],
                 tileSize: 512,
-                zooms: [0, 2, 3, 4, 5, 6, 7, 8, 10, 12, 13, 14],
+                maxZoom: 14,
+                skipZooms: [1, 9, 11]
             },
             "satellite": {
                 type: 'raster',
-                urls: ['http://api.tiles.mapbox.com/v3/aibram.map-vlob92uz/{z}/{x}/{y}.png'],
-                zooms: [0,1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17],
+                url: 'http://api.tiles.mapbox.com/v3/aibram.map-vlob92uz/{z}/{x}/{y}.png',
+                maxZoom: 17
             }
         },
-        maxZoom: 20,
         zoom: 15,
         center: [38.912753, -77.032194],
         style: style_json,