+++ js/llmr.js
@@ -16,10 +16,11 @@ if (typeof window === 'undefined') {
     llmr.StyleDeclaration = require('./style/styledeclaration.js');
 
     llmr.LatLng = require('./geometry/latlng.js');
+    llmr.LatLngBounds = require('./geometry/latlngbounds.js');
     llmr.Point = require('./geometry/point.js');
 
     llmr.Tile = require('./ui/tile.js');
 
-    llmr.Evented = require('./lib/evented.js');
+    llmr.Evented = require('./util/evented.js');
     llmr.util = require('./util/util.js');
 }